+++ src/simulation/simulates.py
@@ -14,48 +14,10 @@ from datetime import datetime, timedelta
 from src.discord.configs import get_simulations_config
 from src.simulation.positions import get_positions
 from src.discord.embeds import discord 
-import json 
-#from loguru import logger
-
-async def send_current_positions_embed(simulator, channel_id, start_ts, end_ts, positions):
-    guild = simulator.discord_bot.get_guild(int(simulator.bot_config.get("discord_id", "")))  # get the guild (server)
-    channel = guild.get_channel(int(channel_id))  # get the channel by ID
-    if channel is None:
-        return
-
-    if isinstance(end_ts, str):
-        end_ts_datetime = datetime.strptime(end_ts, "%Y-%m-%d")
-    else:
-        end_ts_datetime = end_ts
-
-    title_line = f"⌛ Current Positions"
-    
-    description_lines = []
-    description_lines.append('pair | buy_date | buy_price | duration')
-    #description_lines.append('')
-    for position in positions:
-        buy_date_datetime = datetime.strptime(position['buy_date'], "%Y-%m-%d")
-        duration = end_ts_datetime - buy_date_datetime
-        line = f"{position['pair']} | {position['buy_date']} | {position['buy_price']} | {duration.days} days"
-        description_lines.append(line)
-    
-    embed = discord.Embed(
-        title=title_line,
-        description="\n".join(description_lines),
-        color=discord.Color.blue()
-    )
-
-    #embed.set_author(name="Smartswap", icon_url="https://avatars.githubusercontent.com/u/171923264?s=200&v=4")
-    embed.set_footer(text=f"Simulation from {start_ts} to {end_ts}")
-
-    # Check for duplicate messages in the last 10 messages
-    async for message in channel.history(limit=10):
-        if message.author == simulator.discord_bot.user and message.embeds:
-            if message.embeds[0].description == embed.description and message.embeds[0].footer.text == embed.footer.text:
-                return  # Do not send the message if it's the same as the last one
-
-    await channel.send(embed=embed)
-    
+from src.discord.logs.current_positions import send_current_positions_embed
+from src.db.tables import create_funds_table
+from src.discord.logs.position import send_position_embed
+from src.db.simulation import save_simulation_data    
 
 async def simulates(simulator):
     """
@@ -66,7 +28,7 @@ async def simulates(simulator):
     async with aiohttp.ClientSession() as session:
         simulates_config = get_simulations_config()  # get simulations configuration
         for simulation_name, simulation in simulates_config.items():
-            simulator.db_manager.create_funds_table(simulation_name, simulation['positions']['position_%_invest'])
+            await create_funds_table(simulator.db_manager, simulation_name, simulation['positions']['position_%_invest'])
             start_ts_config = datetime.strptime(simulation["api"]["start_ts"], "%Y-%m-%d")
 
             if simulation["api"]["end_ts"]:
@@ -97,7 +59,7 @@ async def simulates(simulator):
 
 
                 # save simulation data
-                simulator.db_manager.save_simulation_data(simulation_name, start_ts_config.strftime("%Y-%m-%d"), end_ts.strftime("%Y-%m-%d"))
+                save_simulation_data(simulator.db_manager, simulation_name, start_ts_config.strftime("%Y-%m-%d"), end_ts.strftime("%Y-%m-%d"))
 
                 if len(positions) > 0:
                     # process each position
@@ -123,7 +85,7 @@ async def simulates(simulator):
                                                                 (pair, buy_date, buy_price, sell_date, sell_price))
                             simulator.db_manager.db_connection.commit()
                             position_id = simulator.db_manager.db_cursor.lastrowid
-                            await simulator.send_position_embed(simulation['discord']['discord_channel_id'], "🚀 Opened Position", discord.Color.brand_green(), position)
+                            await send_position_embed(simulator, simulation['discord']['discord_channel_id'], "🚀 Opened Position", discord.Color.brand_green(), position)
 
 
                         # insert or replace into simulation_positions table
@@ -140,6 +102,6 @@ async def simulates(simulator):
                         positions=positions
                     )
                 else:
-                    await simulator.send_position_embed(simulation['discord']['discord_channel_id'], f"❌ No any position ({start_ts_config}-{end_ts})", discord.Color.red(), {})
+                    await send_position_embed(simulator, simulation['discord']['discord_channel_id'], f"❌ No any position ({start_ts_config}-{end_ts})", discord.Color.red(), {})
 
                 end_ts += timedelta(days=1)  # increment end timestamp by one day
